TR update 2
