+++ python-get-all-files-in-directory/bonus/iterdir.py
