+++ python-get-all-files-in-directory/bonus/scandir.py
